KT-1 ver_4 (double to float output)

+++ Src/searchresult.h
@@ -23,7 +23,7 @@
 
 struct SearchResult {
         bool pathfound;
-        double pathlength; //if path not found, then pathlength=0
+        float pathlength; //if path not found, then pathlength=0
         std::list<Node> lppath; //path as the sequence of adjacent nodes (see above)
                                 //This is a pointer to the list of nodes that is actually created and hadled by Search class,
                                 //so no need to re-create them, delete them etc. It's just a trick to save some memory
